+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/UtxosSpec.hs
@@ -344,36 +344,36 @@ conwayFeaturesPlutusV1V2FailureSpec = do
                 )
         describe "DelegTxCert" $ do
           it "V1" $ do
-            (drepKH, delegatorKH, _spendingKP) <- setupSingleDRep 1_000
+            (drep, delegator, _) <- setupSingleDRep 1_000
             let delegTxCert =
                   DelegTxCert @era
-                    (KeyHashObj delegatorKH)
-                    (DelegVote (DRepCredential $ KeyHashObj drepKH))
+                    delegator
+                    (DelegVote (DRepCredential drep))
             testCertificateNotSupportedV1 delegTxCert
           it "V2" $ do
-            (drepKH, delegatorKH, _spendingKP) <- setupSingleDRep 1_000
+            (drep, delegator, _) <- setupSingleDRep 1_000
             let delegTxCert =
                   DelegTxCert @era
-                    (KeyHashObj delegatorKH)
-                    (DelegVote (DRepCredential $ KeyHashObj drepKH))
+                    delegator
+                    (DelegVote (DRepCredential drep))
             testCertificateNotSupportedV2 delegTxCert
         describe "RegDepositDelegTxCert" $ do
           it "V1" $ do
-            (drepKH, _delegatorKH, _spendingKP) <- setupSingleDRep 1_000
+            (drep, _, _) <- setupSingleDRep 1_000
             unregisteredDelegatorKH <- freshKeyHash
             let regDepositDelegTxCert =
                   RegDepositDelegTxCert @era
                     (KeyHashObj unregisteredDelegatorKH)
-                    (DelegVote (DRepCredential $ KeyHashObj drepKH))
+                    (DelegVote (DRepCredential drep))
                     (Coin 0)
             testCertificateNotSupportedV1 regDepositDelegTxCert
           it "V2" $ do
-            (drepKH, _delegatorKH, _spendingKP) <- setupSingleDRep 1_000
+            (drep, _, _) <- setupSingleDRep 1_000
             unregisteredDelegatorKH <- freshKeyHash
             let regDepositDelegTxCert =
                   RegDepositDelegTxCert @era
                     (KeyHashObj unregisteredDelegatorKH)
-                    (DelegVote (DRepCredential $ KeyHashObj drepKH))
+                    (DelegVote (DRepCredential drep))
                     (Coin 0)
             testCertificateNotSupportedV2 regDepositDelegTxCert
         describe "AuthCommitteeHotKeyTxCert" $ do
@@ -407,21 +407,21 @@ conwayFeaturesPlutusV1V2FailureSpec = do
             testCertificateNotSupportedV2 regDRepTxCert
         describe "UnRegDRepTxCert" $ do
           it "V1" $ do
-            (drepKH, _delegatorKH, _spendingKP) <- setupSingleDRep 1_000
-            let unRegDRepTxCert = UnRegDRepTxCert @era (KeyHashObj drepKH) (Coin 0)
+            (drepKH, _, _) <- setupSingleDRep 1_000
+            let unRegDRepTxCert = UnRegDRepTxCert @era drepKH (Coin 0)
             testCertificateNotSupportedV1 unRegDRepTxCert
           it "V1" $ do
-            (drepKH, _delegatorKH, _spendingKP) <- setupSingleDRep 1_000
-            let unRegDRepTxCert = UnRegDRepTxCert @era (KeyHashObj drepKH) (Coin 0)
+            (drepKH, _, _) <- setupSingleDRep 1_000
+            let unRegDRepTxCert = UnRegDRepTxCert @era drepKH (Coin 0)
             testCertificateNotSupportedV2 unRegDRepTxCert
         describe "UpdateDRepTxCert" $ do
           it "V1" $ do
-            (drepKH, _delegatorKH, _spendingKP) <- setupSingleDRep 1_000
-            let updateDRepTxCert = UpdateDRepTxCert @era (KeyHashObj drepKH) SNothing
+            (drepKH, _, _) <- setupSingleDRep 1_000
+            let updateDRepTxCert = UpdateDRepTxCert @era drepKH SNothing
             testCertificateNotSupportedV1 updateDRepTxCert
           it "V2" $ do
-            (drepKH, _delegatorKH, _spendingKP) <- setupSingleDRep 1_000
-            let updateDRepTxCert = UpdateDRepTxCert @era (KeyHashObj drepKH) SNothing
+            (drepKH, _, _) <- setupSingleDRep 1_000
+            let updateDRepTxCert = UpdateDRepTxCert @era drepKH SNothing
             testCertificateNotSupportedV2 updateDRepTxCert
 
 govPolicySpec ::
@@ -434,15 +434,11 @@ govPolicySpec = do
   describe "Gov policy scripts" $ do
     it "failing native script govPolicy" $ do
       (committeeMember :| _) <- registerInitialCommittee
-      (drepKh, _, _) <- setupSingleDRep 1_000_000
+      (dRep, _, _) <- setupSingleDRep 1_000_000
       scriptHash <- impAddNativeScript $ RequireTimeStart (SlotNo 1)
       anchor <- arbitrary
       void $
-        enactConstitution
-          SNothing
-          (Constitution anchor (SJust scriptHash))
-          (KeyHashObj drepKh)
-          committeeMember
+        enactConstitution SNothing (Constitution anchor (SJust scriptHash)) dRep committeeMember
       rewardAccount <- registerRewardAccount
       pp <- getsNES $ nesEsL . curPParamsEpochStateL
       impAnn "ParameterChange" $ do
@@ -481,14 +477,14 @@ govPolicySpec = do
     it "alwaysSucceeds Plutus govPolicy validates" $ do
       let alwaysSucceedsSh = hashPlutusScript (alwaysSucceeds2 SPlutusV3)
       (committeeMember :| _) <- registerInitialCommittee
-      (drepKh, _, _) <- setupSingleDRep 1_000_000
+      (dRep, _, _) <- setupSingleDRep 1_000_000
       anchor <- arbitrary
       pp <- getsNES $ nesEsL . curPParamsEpochStateL
       void $
         enactConstitution
           SNothing
           (Constitution anchor (SJust alwaysSucceedsSh))
-          (KeyHashObj drepKh)
+          dRep
           committeeMember
       rewardAccount <- registerRewardAccount
 
@@ -519,15 +515,11 @@ govPolicySpec = do
     it "alwaysFails Plutus govPolicy does not validate" $ do
       let alwaysFailsSh = hashPlutusScript (alwaysFails2 SPlutusV3)
       (committeeMember :| _) <- registerInitialCommittee
-      (drepKh, _, _) <- setupSingleDRep 1_000_000
+      (dRep, _, _) <- setupSingleDRep 1_000_000
       anchor <- arbitrary
       pp <- getsNES $ nesEsL . curPParamsEpochStateL
       void $
-        enactConstitution
-          SNothing
-          (Constitution anchor (SJust alwaysFailsSh))
-          (KeyHashObj drepKh)
-          committeeMember
+        enactConstitution SNothing (Constitution anchor (SJust alwaysFailsSh)) dRep committeeMember
 
       rewardAccount <- registerRewardAccount
       impAnn "ParameterChange" $ do
@@ -570,20 +562,20 @@ costModelsSpec =
       modifyPParams $ ppCostModelsL .~ testingCostModels [PlutusV1 .. PlutusV2]
 
       (committeeMember :| _) <- registerInitialCommittee
-      (drepKh, _, _) <- setupSingleDRep 1_000_000
+      (dRep, _, _) <- setupSingleDRep 1_000_000
       anchor <- arbitrary
       govIdConstitution1 <-
-        enactConstitution SNothing (Constitution anchor SNothing) (KeyHashObj drepKh) committeeMember
+        enactConstitution SNothing (Constitution anchor SNothing) dRep committeeMember
       -- propose and enact PlutusV3 Costmodels
       govIdPPUpdate1 <-
-        enactCostModels SNothing (testingCostModels [PlutusV3]) (KeyHashObj drepKh) committeeMember
+        enactCostModels SNothing (testingCostModels [PlutusV3]) dRep committeeMember
 
       let alwaysFailsSh = hashPlutusScript (alwaysFails2 SPlutusV3)
       void $
         enactConstitution
           (SJust (GovPurposeId govIdConstitution1))
           (Constitution anchor (SJust alwaysFailsSh))
-          (KeyHashObj drepKh)
+          dRep
           committeeMember
 
       impAnn "Fail to update V3 Costmodels" $ do
@@ -605,14 +597,14 @@ costModelsSpec =
       modifyPParams $ ppCostModelsL .~ testingCostModels [PlutusV1 .. PlutusV2]
 
       (committeeMember :| _) <- registerInitialCommittee
-      (drepKh, _, _) <- setupSingleDRep 1_000_000
+      (dRep, _, _) <- setupSingleDRep 1_000_000
       anchor <- arbitrary
       let alwaysSucceedsSh = hashPlutusScript (alwaysSucceeds2 SPlutusV3)
       void $
         enactConstitution
           SNothing
           (Constitution anchor (SJust alwaysSucceedsSh))
-          (KeyHashObj drepKh)
+          dRep
           committeeMember
 
       let pparamsUpdate = def & ppuCostModelsL .~ SJust (testingCostModels [PlutusV3])
@@ -624,10 +616,10 @@ costModelsSpec =
       modifyPParams $ ppCostModelsL .~ testingCostModels [PlutusV1 .. PlutusV2]
 
       (committeeMember :| _) <- registerInitialCommittee
-      (drepKh, _, _) <- setupSingleDRep 1_000_000
+      (dRep, _, _) <- setupSingleDRep 1_000_000
       anchor <- arbitrary
       govIdConstitution1 <-
-        enactConstitution SNothing (Constitution anchor SNothing) (KeyHashObj drepKh) committeeMember
+        enactConstitution SNothing (Constitution anchor SNothing) dRep committeeMember
 
       let guessTheNumberSh = hashPlutusScript (guessTheNumber3 SPlutusV3)
 
@@ -639,7 +631,7 @@ costModelsSpec =
         enactCostModels
           SNothing
           (testingCostModels [PlutusV3])
-          (KeyHashObj drepKh)
+          dRep
           committeeMember
 
       let alwaysSucceedsSh = hashPlutusScript (alwaysSucceeds2 SPlutusV3)
@@ -647,7 +639,7 @@ costModelsSpec =
         enactConstitution
           (SJust (GovPurposeId govIdConstitution1))
           (Constitution anchor (SJust alwaysSucceedsSh))
-          (KeyHashObj drepKh)
+          dRep
           committeeMember
 
       impAnn "Minting token succeeds" $ do
@@ -659,7 +651,7 @@ costModelsSpec =
           enactCostModels
             (SJust govIdPPUpdate1)
             (testingCostModels [PlutusV3])
-            (KeyHashObj drepKh)
+            dRep
             committeeMember
 
 txWithPlutus ::